Add ExperienceForm component
Accepts an array of sections and conditionally renders a Sections
component. When a section is added/edited via the corresponding
buttons, the component switches to rendering a form instead.

The form provides inputs for the user to enter details for the
individual section, such as employer, position, and year.
Buttons are included for canceling/saving any changes.